requirements.txt
ppaquette-gym-super-mario
neat-python
numpy
matplotlib
graphviz
# Optional for PPO training/run
stable-baselines3==1.8.0
torch>=1.11
